feat(ui): Add styles -style form [Finishes #166625605] [Finishes ##166625581] [Finishes #166625605] [starts #166625288]
