+++ UI/createad.html
@@ -11,7 +11,7 @@
   <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.8.1/css/all.css" integrity="sha384-50oBUHEmvpQ+1lW4y57PTFmhCaXp0ML5d60M1M7uH2+nqUivzIebhndOJK28anvf" crossorigin="anonymous">
   <link rel="stylesheet" href="css/main.css">
   <link rel="stylesheet" href="css/form.css">
-  <link rel="stylesheet" href="css/createad.css">
+  <link rel="stylesheet" href="css/details_view.css">
  </head>
 
 <body>
@@ -81,6 +81,7 @@
     &copy;2019
   </footer>
   <script src="js/map.js"></script>
+  <script type="module" src="js/details_view.js"></script>
   <script type="module" src="js/createad.js"></script>
   
 </body>
