+++ UI/css/details_view.css
@@ -1,31 +1,41 @@
 #display_property {
-  margin: 20px auto;
-  width:70%;
-  display: grid;
+  display: none;
+  width:70vw;
   grid-template-columns: 2fr 2fr;
   align-items: center;
   justify-content:center;
+  position: relative;
+  transition: display 3s ease 2s;
+  margin: 20px auto;
 }
 
 #display_property div {
   border-radius:10px;
   width:300px;
   height:300px;
+  margin-right:0px;
 }
 
 
 #image_display {
   border-radius:10px;
+  padding:0px;
+ 
 }
-  
-#data_display li {
 
+#data_display {
+  margin-left:0;
+}
+#data_display li {
+font-size:1em;
+padding:0;
+margin:.5em 0;
+text-align: left;
 }
 #data_display {
   color: white;
   display: flex;
   flex-direction: column;
-  list-style-type: square circle; 
 }
 
 #data_display li{
@@ -38,4 +48,14 @@
   width:600px;
   margin: 20px auto;
   transition: display 3s ease;
+}
+
+@media(max-width:800px){
+  #display_property{
+    grid-template-columns:1fr;
+  }
+
+  #data_display {
+    padding-left:0px;
+  }
 }
