+++ UI/user.html
@@ -10,6 +10,7 @@
   
   <link rel="stylesheet" href="css/main.css">
   <link rel="stylesheet" href="css/user.css">
+  <link rel="stylesheet" href="css/details_view.css">
 </head>
 
 <body>
@@ -52,10 +53,21 @@
            Click to see all available Real Estate
          </p>
     </div>
+    <div id="display_property">
+      <div> 
+        <img id="image_display">
+      </div>
+        <ul id="data_display">
+ 
+        </ul>
+    </div>
+    <div id="map">
+    </div>
   </main>
   <footer>
     &copy;2019
   </footer>
+  <script type="module" src="js/details_view.js"></script>
 </body>
 
 </html>
